requirements.txt
absl-py==1.4.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asciitree==0.3.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
fonttools==4.40.0
fqdn==1.5.1
freetype-py==2.4.0
fuzzywuzzy==0.18.0
google-auth==2.19.1
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
grpcio==1.54.2
h5py==3.8.0
hsluv==5.0.3
idna==3.4
igraph==0.10.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
molesq==0.4.0
morphops==0.1.13
navis==1.3.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
ncollpyde==0.19.0
nest-asyncio==1.5.6
networkx==3.1
neuprint-python==0.4.25
notebook==6.5.4
notebook_shim==0.2.3
nptyping==2.5.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
Pint==0.22
platformdirs==3.5.3
plotly==5.15.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.6.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pynrrd==1.0.0
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
rdata==0.9
rdflib==6.3.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
skeletor==1.2.3
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
tenacity==8.2.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
terminado==0.17.1
texttable==1.6.7
threadpoolctl==3.1.0
tinycss2==1.2.1
torch==1.8.0
torch-cluster==1.5.9
torch-geometric==2.0.1
torch-scatter==2.0.8
torch-sparse==0.6.12
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trimesh==3.22.0
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.8.0
uri-template==1.2.0
urllib3==1.26.16
vispy==0.13.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.3
Werkzeug==2.3.6
widgetsnbextension==4.0.7
xarray==2023.5.0
yacs==0.1.8
zipp==3.15.0
